Moving another agent test to legacy suite
The following moves the model runs instance poller test from the
machine test to legacy one.
